requirements.txt
discord
mss
requests
pyinstaller
pyarmor
sqlite3
win32gui
pycryptodome
pywin32